ld enigma as one static binary... no more libenigma*.so* files needed..., small combobox fix, add textinput control ( TODO: changed characters for dbox2 ), fix warnings, add multi sortable user bouquets ( use the red button in the service selector to create a new user Bouquet, or to copy a complete bouquet from a dvb bouquet to user bouquet list), to chage to user bouquets use the blue key in serice selector.. this toggles between DVB -Bouquets, Playlist and user Bouquets,  add quick zap funktion between to last channels... ( on the 0 ), service numbers < 200 are readed from the first userbouquet.. move your favourite list to first pos..., fix bugs in transponderscan, remove German Free Bouquet, please create a userbouquet... , fix segfault in software update menu

+++ tuxbox/enigma/src/enigma_event.cpp
@@ -144,8 +144,8 @@ eEventDisplay::eEventDisplay(eString service, const ePtrList<EITEvent>* e, EITEv
 	// try to recalc long description label... ( no broken text lines.. )
 	float lineheight=fontRenderClass::getInstance()->getLineHeight( long_description->getFont() );
 	int lines = descr->getSize().height() / (int)lineheight;
-	int newheight = lines * (int)lineheight + (round(lineheight) - (int)lineheight);
-	descr->resize( eSize( descr->getSize().width(), newheight + lineheight/6 ) );
+	int newheight = lines * (int)lineheight + (int)(round(lineheight) - (int)lineheight);
+	descr->resize( eSize( descr->getSize().width(), newheight + (int)lineheight/6 ) );
 	long_description->resize(eSize(descr->getSize().width(), descr->getSize().height()*4));
 
 	if (e)
